Register and support opening irc URL schemes
The irc URL scheme typically refer a channel on some
server. This patch register Polari as
handler of the x-scheme-handler/irc mimetype. If Polari
is the default handler, browsers can activate Polari
via xdg-open/d-bus activation. Polari will attempt to
parse the link and if it fails, it will emit a
notification with an error.

https://bugzilla.gnome.org/show_bug.cgi?id=728593